requirements.txt
paho-mqtt
emoji
python-dotenv
python-telegram-bot
